% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dp.R
\name{read_dp}
\alias{read_dp}
\title{Read Spectrum .DP file}
\usage{
read_dp(pjnz)
}
\arguments{
\item{pjnz}{file path to Spectrum PJNZ file}
}
\value{
Named list of PJNZ data.
}
\description{
This function returns the Spectrum .DP file as a named list of vectors.
}
\examples{
pjnz <- system.file(
  "pjnz", "bwa_aim-adult-art-no-special-elig_v6.13_2022-04-18.PJNZ",
  package = "pjnz"
)
dp <- read_dp(pjnz)
}
